detection/requirements.txt
absl-py==2.1.0
albucore==0.0.23
albumentations==1.4.24
annotated-types==0.7.0
anyio==4.7.0
astor==0.8.1
astunparse==1.6.3
Automat==24.8.1
blinker==1.9.0
cattrs==24.1.2
chex==0.1.88
coloredlogs==15.0.1
Comet==3.1.0
constantly==23.10.4
contourpy==1.3.1
coremltools==8.1
cycler==0.12.1
dnspython==2.7.0
etils==1.11.0
expecttest==0.2.1
Flask==3.1.0
flatbuffers==24.12.23
flax==0.10.2
fonttools==4.55.3
fsspec==2024.9.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.1
h11==0.14.0
h5py==3.12.1
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
humanfriendly==10.0
humanize==4.11.0
hyperlink==21.0.0
hypothesis==6.115.3
incremental==24.7.2
itsdangerous==2.2.0
jax==0.4.38
jaxlib==0.4.38
joblib==1.4.2
keras==3.7.0
kiwisolver==1.4.8
libclang==18.1.1
lintrunner==0.12.5
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.0
namex==0.0.8
ncnn==1.0.20241226
nest-asyncio==1.6.0
networkx==3.4.1
ninja==1.11.1.1
numpy==1.23.5
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnx==1.17.0
onnx-graphsurgeon==0.5.2
onnx2tf==1.26.3
onnxruntime==1.20.1
onnxslim==0.1.45
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openvino==2024.6.0
openvino-telemetry==2024.5.0
opt-einsum==3.3.0
optax==0.2.4
optree==0.13.0
orbax-checkpoint==0.11.0
packaging==23.2
paddlepaddle==2.6.2
pandas==2.2.3
pillow==10.2.0
portalocker==3.1.1
protobuf==3.20.3
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
pyaml==25.1.0
pybind11==2.13.6
pycocotools==2.0.8
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-etcd==0.4.5
rich==13.9.4
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
simplejson==3.19.3
simsimd==6.2.1
sng4onnx==1.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
stringzilla==3.11.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow_decision_forests==1.11.0
tensorflowjs==4.22.0
tensorrt_cu12==10.7.0
tensorstore==0.1.71
termcolor==2.5.0
tf_keras==2.18.0
tflite-support==0.4.4
threadpoolctl==3.5.0
toolz==1.0.0
torch==2.5.0+cu124
torchaudio==2.5.0+cu124
torchelastic==0.2.2
torchvision==0.20.0+cu124
triton==3.1.0
Twisted==24.11.0
types-dataclasses==0.6.6
tzdata==2024.2
ultralytics-thop==2.0.13
uv==0.5.14
Werkzeug==3.1.3
wrapt==1.17.0
wurlitzer==3.1.1
x2paddle==1.6.0
ydf==0.9.0
zope.interface==7.2
zstandard==0.23.0
